Moved TagInfo component into components folder
